Core-Integrator-Sprint-1.1-/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
python-multipart>=0.0.6
pytest>=8.0.0
pytest-asyncio>=0.23.0
requests>=2.0.0
PyNaCl>=1.5.0
pymongo>=4.0.0